refactor(convention): remove dead code for convention and siret search
