+++ packages/code-du-travail-api/src/server/routes/__tests__/idcc.spec.js
@@ -13,19 +13,8 @@ test("return idcc results for boulangerie", async () => {
   expect(response.body.hits).toMatchSnapshot();
 });
 
-test("return idcc results for num 843", async () => {
-  const response = await request(app.callback()).get(`/api/v1/idcc/843`);
-  expect(response.status).toBe(200);
-  expect(response.body).toMatchSnapshot();
-});
-
 test("return idcc results for pati", async () => {
   const response = await request(app.callback()).get(`/api/v1/idcc?q=pati`);
   expect(response.status).toBe(200);
   expect(response.body.hits).toMatchSnapshot();
-});
-
-test("returns 404 when IDCC with num not found", async () => {
-  const response = await request(app.callback()).get(`/api/v1/idcc/8434`);
-  expect(response.status).toBe(404);
 });
