+++ packages/code-du-travail-api/src/server/routes/idcc/index.js
@@ -3,7 +3,6 @@ const API_BASE_URL = require("../v1.prefix");
 
 const elasticsearchClient = require("../../conf/elasticsearch.js");
 const getIdccBody = require("./idcc.elastic");
-const getIdccByNumBody = require("./idccByNum.elastic");
 
 const index =
   process.env.ELASTICSEARCH_DOCUMENT_INDEX || "code_du_travail_numerique";
@@ -26,27 +25,6 @@ router.get("/idcc", async ctx => {
 
   const response = await elasticsearchClient.search({ index, body });
   ctx.body = response.body;
-});
-
-/**
- * Return the convention collective object that matches the given num
- *
- * @example
- * http://localhost:1337/api/v1/idcc/200
- *
- * @param {string} :num the searched IDCC number
- * @returns {Object} one IDCC object.
- */
-router.get("/idcc/:num", async ctx => {
-  const body = getIdccByNumBody({ query: ctx.params.num });
-
-  const response = await elasticsearchClient.search({ index, body });
-
-  if (response.body.hits.total.value === 0) {
-    ctx.throw(404, `no IDCC found for num ${ctx.params.num}`);
-  }
-
-  ctx.body = { ...response.body.hits.hits[0] };
 });
 
 module.exports = router;
