+++ packages/code-du-travail-frontend/src/conventions/Search/Form.js
@@ -5,11 +5,7 @@ import { theme, Container, Table, Tag } from "@cdt/ui-old";
 
 import SearchCC from "./SearchCC";
 
-// normalize idcc nums
-const formatIdCc = num => {
-  while (num && num.length < 4) num = "0" + num;
-  return num;
-};
+
 
 // link to a CC
 const CC = ({ id, num, title }) => {
@@ -17,7 +13,7 @@ const CC = ({ id, num, title }) => {
     <Box>
       <Flex>
         <Tag title="Numéro de convention collective" variant="info">
-          IDCC {formatIdCc(num)}
+          IDCC {`0000${num}`.slice(-4)}
         </Tag>
         <Spacer />
         <CCLink
