+++ packages/code-du-travail-frontend/src/conventions/services.js
@@ -23,7 +23,25 @@ const getConventionTextesMemo = memoizee(
   }
 );
 
+function searchIdcc(query) {
+  const url = `${API_URL}/idcc?q=${encodeURIComponent(
+    query.replace(/ /g, "")
+  )}`;
+
+  return fetch(url).then(response => {
+    if (response.ok) {
+      return response.json().then(results => results.hits.hits);
+    }
+    throw new Error("Un problème est survenu.");
+  });
+}
+
+// memoize search results
+const searchIdccMemo = memoizee(query => searchIdcc(query), { promise: true });
+
 export {
   getConventionTextesMemo as getConventionTextes,
-  getConventionTextes as _getConventionTextes
+  getConventionTextes as _getConventionTextes,
+  searchIdccMemo as searchIdcc,
+  searchIdcc as _searchIdcc
 };
